ArticleList 컴포넌트 추가, store 네이밍 수정, 타입 분리
