+++ src/@types/Article.ts
@@ -1,0 +1,4 @@
+import { components } from '.';
+
+export type Article = components['schemas']['Article'];
+export type MultipleArticlesResponse = components['schemas']['MultipleArticlesResponse'];
